chore: cleanup main file, move statusline in module
